arch: split iterator into multiple files
